[fix] updated readme, updated interface
